Adds dungeon room/corridor generation, pathfinding and player movement

+++ TOGA-Final/ecs.h
@@ -4,17 +4,25 @@
 #include "component_manager.h"
 #include "system_manager.h"
 
+#include "components.h"
+
 // Static class that acts as a central location to access ECS related functions
 class ECS {
 public:
+	static std::unique_ptr<EntityManager> entityManager;
+	static std::unique_ptr<ComponentManager> componentManager;
+	static std::unique_ptr<SystemManager> systemManager;
+
 	static Entity createEntity() {
 		return entityManager->create();
 	}
 
 	static void destroyEntity(Entity entity) {
-		entityManager->destroy(entity);
-		componentManager->onEntityDestroyed(entity);
-		systemManager->onEntityDestroyed(entity);
+		if (entityManager->isAlive(entity)) {
+			entityManager->destroy(entity);
+			componentManager->onEntityDestroyed(entity);
+			systemManager->onEntityDestroyed(entity);
+		}
 	}
 
 	template<typename Component>
@@ -24,20 +32,20 @@ public:
 
 	// Maybe move these into ComponentManager
 	template<typename Component>
-	static void addComponent(Entity& entity, Component component) {
+	static void addComponent(Entity entity, Component component) {
 		// Adds the component to the vector of components
 		componentManager->addComponent<Component>(entity, component);
 		// Sets the bit at the index of the component's unique id to true
-		entity.signature.set(componentManager->getComponentId<Component>(), true);
+		entityManager->getSignature(entity).set(componentManager->getComponentId<Component>(), true);
 		// Update systems to add entities if needed
-		systemManager->onEntitySignatureChanged(entity);
+		systemManager->onEntitySignatureChanged(entity, entityManager->getSignature(entity));
 	}
 
 	template<typename Component>
 	static void removeComponent(Entity entity) {
 		componentManager->removeComponent<Component>(entity);
-		entity.signature.set(componentManager->getComponentId<Component>(), false);
-		systemManager->onEntitySignatureChanged(entity);
+		entityManager->getSignature(entity).set(componentManager->getComponentId<Component>(), false);
+		systemManager->onEntitySignatureChanged(entity, entityManager->getSignature(entity));
 	}
 
 	template<typename Component>
@@ -51,17 +59,7 @@ public:
 	}
 
 	template<typename System>
-	static std::shared_ptr<System> registerSystem(Signature signature) {
-		return systemManager->registerSystem<System>(signature);
+	static std::shared_ptr<System> registerSystem() {
+		return systemManager->registerSystem<System>();
 	}
-
-	//template<typename System>
-	//static void setSystemSignature(Signature signature) {
-	//	systemManager->setSignature<System>(signature);
-	//}
-
-private:
-	static std::unique_ptr<EntityManager> entityManager;
-	static std::unique_ptr<ComponentManager> componentManager;
-	static std::unique_ptr<SystemManager> systemManager;
 };
